requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.0.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.109.2
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-language==2.13.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jiter==0.5.0
openai==1.40.6
passlib==1.7.4
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.27.1